fix tests

+++ spgl1/spgl1.py
@@ -554,7 +554,7 @@ def _spg_line_curvy(x, g, fmax, A, b, project, weights, tau):
     while 1:
         # Evaluate trial point and function value.
         start_time_project = time.time()
-        xnew = project(x - step*scale*g, weights, tau)
+        xnew = project(x - (step*scale*g).astype(g.dtype), weights, tau)
         timeproject += time.time() - start_time_project
         start_time_matprod = time.time()
         rnew = b - A.matvec(xnew)
